Store content-disposition and content-type in GCS when forced
Forcing these in the service_url when serving the file works fine for S3
and Azure, since these services include params in the signature.
However, GCS specifically excludes response-content-disposition and
response-content-type from the signature, which means an attacker can
modify these and have files that should be served as binary served as
they wish. The only way this can be forced in GCS is by storing
content-disposition and content-type in the object metadata.