Remove deprecation of `query_constraints:` association option, raise instead
Deprecation was added in https://github.com/rails/rails/pull/51571 with the intention
to change the behavior of `query_constraints:` option on associations.

The change is still intended to be implemented but was delayed to a future Rails version.
Meanwhile we would like applications to stop using `query_constraints:` option on associations.